requirements.txt
fastapi==0.111.0
numpy==1.26.4
pillow==10.3.0
tensorflow==2.16.1
uvicorn==0.30.1
